feat: lista de pedidos y correcciones
