+++ app/Http/Controllers/Admin/SaleController.php
@@ -67,7 +67,7 @@ class SaleController extends Controller
 
     public function updateTracking(Sale $sale, Request $request)
     {
-        $this->updateStatus($sale, $request->status, null);
+        $this->updateStatus($sale, $request->status);
         return back()->with('success', 'Estado actualizado');
     }
     /**
@@ -75,45 +75,7 @@ class SaleController extends Controller
      */
     public function store(Request $request)
     {
-        try {
-            $validate = $request->validate([
-                'user_id' => ['required', 'integer', 'exists:users,id'],
-                'transaction_number' => ['required', 'string'],
-                'payment_type' => ['required', 'integer'],
-                'items' => ['required', 'array'],
-                'date_generated' => ['required', 'date_format:Y-m-d H:i:s'],
-            ]);
-            DB::beginTransaction();
-            $validate['status'] = 1;
-            $validate['amount_paid'] = 0;
-            $validate['tracking'] = json_encode([$validate['date_generated']]);
-            $sale = Sale::create($validate);
-            $this->updateStatus($sale,1);
-            $amount_paid = 0;
-            foreach ($validate['items'] as $item) {
-                $product = Product::find($item['id']);
-                if ($product) {
-                    $sale->sale_details()->create([
-                        'product_id' => $item['id'],
-                        'amount' => $item['amount'],
-                    ]);
-                    $amount_paid += $product->price * $item['amount'];
-                }
-            }
-            $sale->amount_paid = $amount_paid;
-            $sale->save();
-            DB::commit();
-            $this->message = 'Se ha generado tu pedido correctamente';
-            $this->status_code = 200;
-        } catch (\Exception $exception) {
-            DB::rollBack();
-            $this->message = $exception->getMessage();
-            $this->status_code = 500;
-        }
-        $this->response = [
-            'message' => $this->message
-        ];
-        return response()->json($this->response, $this->status_code);
+        //
     }
 
     /**
