+++ routes/api.php
@@ -1,7 +1,5 @@
 <?php
 
-use App\Http\Controllers\Admin\SaleController;
-use App\Http\Controllers\Api\ProductController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -20,9 +18,4 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::middleware('auth')->group(function () {
 
-});
-Route::post('create/sale', [SaleController::class, 'store']);
-
-Route::get('products/{product}/image-gallery', [ProductController::class, 'imageGallery'])->name('api.products.imageGallery');
